feat: typescript deme remove @types/egg dependency
