+++ hackernews-async-ts/app/controller/index.d.ts
@@ -1,0 +1,6 @@
+import NewsController from './news';
+declare module 'egg' {
+  export interface IController {
+    news: NewsController;
+  }
+}
