+++ hackernews-async-ts/app/service/index.d.ts
@@ -1,0 +1,6 @@
+import HackerNews from './HackerNews';
+declare module 'egg' {
+  export interface IService {
+    hackerNews: HackerNews;
+  }
+}
